fix(cli): fix yarn berry detection
closes #11495
