+++ .changes/cli-yarn-berry.md
@@ -1,0 +1,6 @@
+---
+"tauri-cli": "patch:bug"
+"@tauri-apps/cli": "patch:bug"
+---
+
+Fix detecting yarn berry (v2 and higher) in various tauri cli commands.
