+++ crates/tauri-cli/src/info/env_nodejs.rs
@@ -5,23 +5,7 @@
 use super::{ActionResult, SectionItem, VersionMetadata};
 use colored::Colorize;
 
-use crate::helpers::cross_command;
-
-pub fn manager_version(package_manager: &str) -> Option<String> {
-  cross_command(package_manager)
-    .arg("-v")
-    .output()
-    .map(|o| {
-      if o.status.success() {
-        let v = String::from_utf8_lossy(o.stdout.as_slice()).to_string();
-        Some(v.split('\n').next().unwrap().to_string())
-      } else {
-        None
-      }
-    })
-    .ok()
-    .unwrap_or_default()
-}
+use crate::helpers::{cross_command, npm::manager_version};
 
 pub fn items(metadata: &VersionMetadata) -> Vec<SectionItem> {
   let node_target_ver = metadata.js_cli.node.replace(">= ", "");
